refactor: new components to represent bulk enrollment warning, and button
